Remove plural scaffold instructions from git import
