+++ assets/src/components/cd/PrepareGitStep.tsx
@@ -1,36 +1,8 @@
-import { Code } from '@pluralsh/design-system'
 import { useTheme } from 'styled-components'
 
 import { CD_QUICKSTART_LINK } from 'routes/cdRoutesConsts'
 
 import { StepBody, StepH, StepLink } from './ModalAlt'
-
-const scaffoldTabs = [
-  {
-    key: 'nodejs',
-    label: 'Node.js',
-    language: 'sh',
-    content: `plural scaffold --type nodejs --name <my-service>`,
-  },
-  {
-    key: 'rails',
-    label: 'Rails',
-    language: 'sh',
-    content: `plural scaffold --type rails --name <my-service>`,
-  },
-  {
-    key: 'springboot',
-    label: 'Spring boot',
-    language: 'sh',
-    content: `plural scaffold --type springboot --name <my-service>`,
-  },
-  {
-    key: 'django',
-    label: 'Django',
-    language: 'sh',
-    content: `plural scaffold --type django --name <my-service>`,
-  },
-]
 
 export function PrepareGitStep() {
   const theme = useTheme()
@@ -52,8 +24,7 @@ export function PrepareGitStep() {
       >
         <StepH>Step 1. Prepare your Git repository</StepH>
         <StepBody>
-          Need some help to Git ready? Use a plural scaffold to get started or
-          read our{' '}
+          Need some help to Git ready? Read our{' '}
           <StepLink
             href={CD_QUICKSTART_LINK}
             target="_blank"
@@ -63,7 +34,6 @@ export function PrepareGitStep() {
           .
         </StepBody>
       </div>
-      <Code tabs={scaffoldTabs} />
     </div>
   )
 }
